+++ saleor/static/dashboard-next/customers/views/CustomerList.tsx
@@ -3,16 +3,14 @@ import * as React from "react";
 import Navigator from "../../components/Navigator";
 import { createPaginationState, Paginator } from "../../components/Paginator";
 import { maybe } from "../../misc";
+import { Pagination } from "../../types";
 import CustomerListPage from "../components/CustomerListPage";
 import { TypedCustomerListQuery } from "../queries";
 import { customerAddUrl, customerUrl } from "../urls";
 
 const PAGINATE_BY = 20;
 
-export type CustomerListQueryParams = Partial<{
-  after: string;
-  before: string;
-}>;
+export type CustomerListQueryParams = Pagination;
 
 interface CustomerListProps {
   params: CustomerListQueryParams;